requirements.txt
Authlib==1.3.2
aws-lambda-wsgi==0.0.6
aws-wsgi==0.2.7
blinker==1.9.0
Bootstrap-Flask==2.4.1
cachelib==0.13.0
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
click==8.1.7
colorama==0.4.6
cryptography==43.0.3
dnspython==2.7.0
email_validator==2.2.0
Flask==3.1.0
Flask-Login==0.6.3
Flask-Mail==0.10.0
Flask-OAuthlib==0.9.6
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.2
greenlet==3.1.1
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==3.0.2
oauthlib==2.1.0
psycopg2==2.9.10
pycparser==2.22
python-dotenv==1.0.1
requests==2.32.3
requests-oauthlib==1.1.0
SQLAlchemy==2.0.36
typing_extensions==4.12.2
urllib3==2.2.3
Werkzeug==3.1.3
WTForms==3.2.1
